+++ internal/rules/builtin/builtin.go
@@ -51,6 +51,9 @@ func builtins() []rules.Rule {
 		claudeThirdPartyPluginEnabled{},
 		// v0.2.0-alpha.3
 		mcpUnauthRemoteURL{},
+		// v0.2.0-alpha.4
+		cursorAllowlistTooBroad{},
+		cursorMCPWildcard{},
 	}
 }